requirements.txt
attrs==25.1.0
awscli==1.37.21
boto3==1.36.21
botocore==1.36.21
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
dnspython==2.7.0
docutils==0.16
durationpy==0.9
eks_token==0.3.0
google-auth==2.38.0
h11==0.14.0
idna==3.10
iniconfig==2.0.0
jmespath==1.0.1
kubernetes==32.0.0
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.2
pluggy==1.5.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PySocks==1.7.1
pytest==8.3.4
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.7.2
s3transfer==0.11.2
selenium==4.28.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
